Module runner::diffs

Structs

ModifiedContent

Functions

add_affected_crates
Checks if crate_toml_path has a direct or indirect dependency to any of the affected crates in affected_crates_toml_path. If so, adds crate_toml_path and any of its affected dependencies to affected_crates_toml_path.

add_affected_protos
Adds proto_path to the list of affected_protos if proto_path or any of its imported protos imports any of the proto files in affected_protos.

affected_protos
List of paths to all .proto files affected by the recent changes.

all_affected_crates
Path to the Cargo.toml files for all crates that are either directly modified or have a dependency to a modified crate.

crates_affected_by_protos
Returns the paths to Cargo.toml files of crates affected by the changed proto files.

directly_modified_crates
Returns the list of paths to Cargo.toml files for all crates in which at least one file is modified.

find_crate_toml_file

get_local_dependencies
Returns paths to Cargo.toml files of local crates (crates belonging to the Oak repo) that the given crate has a dependency to. Converts the relative dependency paths in Cargo.toml into paths relative to the repo's root.

imported_proto_files
Returns paths to all .proto files that proto_file_path imports.

modified_files