Function runner::diffs::add_affected_crates

fn add_affected_crates(
    crate_toml_path: PathBuf, 
    affected_crates_toml_path: &mut HashMap<String, bool>
)

Checks if crate_toml_path has a direct or indirect dependency to any of the affected crates in affected_crates_toml_path. If so, adds crate_toml_path and any of its affected dependencies to affected_crates_toml_path.

Keys in affected_crates_toml_path are paths to visited Cargo.toml files. If the value associated with a key is true, then the corresponding crate is affected by the changes.