runner

Module exp

Functions

__quantum__qis__exp__adj⚠

QIR API for applying an adjoint exponential of a multi-qubit rotation about the given Pauli axes with the given angle and qubits.

__quantum__qis__exp__body⚠

QIR API for applying an exponential of a multi-qubit rotation about the given Pauli axes with the given angle and qubits.

__quantum__qis__exp__ctl⚠

QIR API for applying an exponential of a multi-qubit rotation about the given Pauli axes with the given angle and qubits.

__quantum__qis__exp__ctladj⚠

QIR API for applying an exponential of a multi-qubit rotation about the given Pauli axes with the given angle and qubits.