Function runner::internal::run_step

pub fn run_step<'life0, 'async_recursion>(
    context: &'life0 Context, 
    step: Step, 
    run_status: Status
) -> Pin<Box<dyn Future<Output = StepResult> + Send + 'async_recursion>> where
    'life0: 'async_recursion, 

Run the provided step, printing out information about the execution, and returning a set of status results from the single or multiple steps that were executed.