Trait runner::internal::Running

pub trait Running: Send {
#[must_use]    fn result<'async_trait>(
        self: Box<Self>
    ) -> Pin<Box<dyn Future<Output = SingleStatusResult> + Send + 'async_trait>>
    where
        Self: 'async_trait;

    fn kill(&mut self) { ... }

    fn stdout(&mut self) -> Box<dyn AsyncRead + Send + Unpin> { ... }

    fn stderr(&mut self) -> Box<dyn AsyncRead + Send + Unpin> { ... }
}

A task that is currently running asynchronously.

Required methods

#[must_use]fn result<'async_trait>(
self: Box<Self>
) -> Pin<Box<dyn Future<Output = SingleStatusResult> + Send + 'async_trait>> where
Self: 'async_trait,

Returns the final result of the task, upon spontaneous termination.

Loading content...
Provided methods

fn kill(&mut self)

Attempts to kill the running task.

fn stdout(&mut self) -> Box<dyn AsyncRead + Send + Unpin>

Returns an AsyncRead object to stream stdout logs from the task.

fn stderr(&mut self) -> Box<dyn AsyncRead + Send + Unpin>

Returns an AsyncRead object to stream stderr logs from the task.

Loading content...