Module runner::runner

Structs

ClientRelease

Data type assumed to be used by the parachain to store the client release. If this type is different from the on-chain one, decoding will fail.

DownloadedRelease

Wrapper around ClientRelease, which includes details for running the executable.

Runner

Per-network manager of the client executable

Enums

ClientType

Type of the client to run. Also used as the name of the downloaded executable.

Constants

BLOCK_TIME

Parachain block time

BRANCHES 🔒

CURRENT_RELEASES_STORAGE_ITEM

Storage item in the Pallet where the client release is assumed to be stored

PARACHAIN_MODULE

Pallet in the parachain where the client release is assumed to be stored

RETRY_INTERVAL

Waiting interval used by the retry utilities: One minute

RETRY_MULTIPLIER

Multiplier for the interval in retry utilities: Constant interval retry

RETRY_TIMEOUT

Timeout used by the retry utilities: One minute

Traits

RunnerExt

StorageReader

Functions

custom_retry_config

retry_with_log

retry_with_log_async

sha256sum 🔒

subxt_api

Opaque Types

🔒

🔒

🔒

🔒

🔒

🔒

🔒