Trait runner::runner::RunnerExt

pub trait RunnerExt {

    // Required methods
    fn subxt_api(&self) -> &OnlineClient<PolkadotConfig>;
    fn client_args(&self) -> &Vec<String> ;
    fn child_proc(&mut self) -> &mut Option<Child>;
    fn set_child_proc(&mut self, child_proc: Option<Child>);
    fn downloaded_release(&self) -> &Option<DownloadedRelease>;
    fn set_downloaded_release(
        &mut self,
        downloaded_release: Option<DownloadedRelease>
    );
    fn download_path(&self) -> &PathBuf;
    fn parachain_url(&self) -> String;
    fn client_type(&self) -> ClientType;
    fn try_get_release<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<ClientRelease>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download_binary<'life0, 'async_trait>(
        &'life0 mut self,
        release: ClientRelease
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_bin_path(&self, uri: &str) -> Result<(String, PathBuf), Error>;
    fn delete_downloaded_release(&mut self) -> Result<(), Error>;
    fn run_binary(&mut self) -> Result<(), Error>;
    fn terminate_proc_and_wait(&mut self) -> Result<(), Error>;
    fn get_request_bytes<'life0, 'async_trait>(
        &'life0 self,
        url: String
    ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn auto_update(&mut self) -> BoxFuture<'_, Result<(), Error>>;
    fn check_child_proc_alive(&mut self) -> Result<bool, Error>;
    fn maybe_restart_client(&mut self) -> Result<(), Error>;
    fn try_load_downloaded_binary(
        &mut self,
        release: &ClientRelease
    ) -> Result<(), Error>;

}

Required Methods

fn subxt_api(&self) -> &OnlineClient<PolkadotConfig>

fn client_args(&self) -> &Vec<String>

fn child_proc(&mut self) -> &mut Option<Child>

fn set_child_proc(&mut self, child_proc: Option<Child>)

fn downloaded_release(&self) -> &Option<DownloadedRelease>

fn set_downloaded_release( &mut self, downloaded_release: Option<DownloadedRelease> )

fn download_path(&self) -> &PathBuf

fn parachain_url(&self) -> String

fn client_type(&self) -> ClientType

fn try_get_release<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Option<ClientRelease>, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Read the current client release from the parachain, retrying for RETRY_TIMEOUT if there is a network error.

fn download_binary<'life0, 'async_trait>( &'life0 mut self, release: ClientRelease ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Download the client binary and make it executable, retrying for RETRY_TIMEOUT if there is a network error.

fn get_bin_path(&self, uri: &str) -> Result<(String, PathBuf), Error>

Convert a release URI (e.g. a GitHub link) to an executable name and OS path (after download)

fn delete_downloaded_release(&mut self) -> Result<(), Error>

Remove downloaded release from the file system. This is only supposed to occur after the client process has been killed. In case of failure, removing is retried for RETRY_TIMEOUT.

fn run_binary(&mut self) -> Result<(), Error>

Spawn a the client as a child process with the CLI arguments set in the Runner, retrying for RETRY_TIMEOUT.

fn terminate_proc_and_wait(&mut self) -> Result<(), Error>

Send a SIGTERM to the child process (via the underlying kill system call). If kill returns an error code, the operation is retried for RETRY_TIMEOUT.

fn get_request_bytes<'life0, 'async_trait>( &'life0 self, url: String ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the client release executable, as Bytes

fn auto_update(&mut self) -> BoxFuture<'_, Result<(), Error>>

Main loop, checks the parachain for new releases and updates the client accordingly.

fn check_child_proc_alive(&mut self) -> Result<bool, Error>

Returns whether the child is alive and sets the runner.child field to None if not.

fn maybe_restart_client(&mut self) -> Result<(), Error>

If the child process crashed, start it again

fn try_load_downloaded_binary( &mut self, release: &ClientRelease ) -> Result<(), Error>

If a client binary exists on disk, load it.