Module runner::internal

Structs

Cmd
A single command that implements the Runnable trait.

Context
Encapsulates all the local state relative to a step, and is propagated to child steps.

Opt

SingleStatusResult

Enums

StatusResultValue
The outcome of an individual step of execution.

Step
An execution step, which may be a single (dynamically dispatched) Runnable, or a collection of sub-steps.

Traits

Runnable
A runnable task which produces a single result.

Functions

cmd
Convenience constructor for a boxed Cmd.

run_step
Run the provided step, printing out information about the execution, and returning a set of status results from the single or multiple steps that were executed.

values_to_string