Function runner::internal::run_step

pub fn run_step(context: &Context, step: &Step) -> HashSet<StatusResultValue>

Run the provided step, printing out information about the execution, and returning a set of status results from the single or multiple steps that were executed.